refactor: extract user profile operations from AuthProvider
- Move updateUsername and checkUsernameAvailable to new useUserProfile hook
- Clean AuthProvider interface to focus only on authentication concerns
- Update UsernameForm to use useUserProfile instead of AuthProvider
- AuthProvider now only handles: user state, sign in/out, wallet connection
- Better separation of concerns: auth vs user profile operations

This makes AuthProvider focused on pure authentication while
user profile operations are handled by dedicated domain hooks.
