+++ app/providers/AuthProvider.tsx
@@ -12,8 +12,7 @@ interface AuthContextType {
   error: string | null;
   signInWithWallet: () => Promise<void>;
   signOut: () => Promise<void>;
-  updateUsername: (username: string) => Promise<void>;
-  checkUsernameAvailable: (username: string) => Promise<boolean>;
+  setUser: (user: User | null) => void;
 }
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
@@ -27,7 +26,6 @@ export function AuthProvider({ children }: AuthProviderProps) {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const { publicKey, connected } = useWallet();
-  const { verifyUsernameChange } = useUserOperations();
   const supabase = createSupabaseBrowserClient();
 
   // Check if user exists or create new user
@@ -101,79 +99,6 @@ export function AuthProvider({ children }: AuthProviderProps) {
     }
   };
 
-  // Update username with wallet signature verification
-  const updateUsername = async (username: string) => {
-    if (!user) {
-      throw new Error('User not authenticated');
-    }
-
-    try {
-      setError(null);
-      
-      // Trim and validate username
-      const trimmedUsername = username.trim();
-      if (!trimmedUsername) {
-        throw new Error('Username cannot be empty');
-      }
-
-      // Check if this is already their current username
-      if (user.username && user.username.toLowerCase() === trimmedUsername.toLowerCase()) {
-        throw new Error('This is already your current username');
-      }
-
-      // Check if username is available
-      const isAvailable = await checkUsernameAvailable(trimmedUsername);
-      if (!isAvailable) {
-        throw new Error('Username is already taken');
-      }
-
-      // Verify wallet ownership with signature
-      await verifyUsernameChange(trimmedUsername);
-
-      // Update username in database
-      const { data: updatedUser, error: updateError } = await supabase
-        .from('users')
-        .update({ username: trimmedUsername })
-        .eq('id', user.id)
-        .select()
-        .single();
-
-      if (updateError) {
-        throw updateError;
-      }
-
-      setUser(updatedUser);
-    } catch (err) {
-      console.error('Error updating username:', err);
-      const errorMessage = err instanceof Error ? err.message : 'Failed to update username';
-      setError(errorMessage);
-      throw new Error(errorMessage);
-    }
-  };
-
-  // Check if username is available
-  const checkUsernameAvailable = async (username: string): Promise<boolean> => {
-    try {
-      const trimmedUsername = username.trim();
-      if (!trimmedUsername) return false;
-
-      const { data, error } = await supabase
-        .from('users')
-        .select('id')
-        .ilike('username', trimmedUsername)
-        .neq('id', user?.id || '');
-
-      if (error) {
-        console.error('Error checking username availability:', error);
-        return false;
-      }
-
-      return data.length === 0;
-    } catch (err) {
-      console.error('Error checking username availability:', err);
-      return false;
-    }
-  };
 
   // Auto-authenticate when wallet connects/disconnects or switches
   useEffect(() => {
@@ -210,8 +135,7 @@ export function AuthProvider({ children }: AuthProviderProps) {
     error,
     signInWithWallet,
     signOut,
-    updateUsername,
-    checkUsernameAvailable,
+    setUser,
   };
 
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
